separate jupyterhublogin tests
